pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reconx"
version = "1.0.0"
description = "Fast security reconnaissance tool — port scan, SSL, headers, DNS, subdomains, paths, WHOIS"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.4.0",
    "python-whois>=0.8.0",
    "requests>=2.31.0",
    "reportlab>=4.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "urllib3>=2.0.0",
]

[project.scripts]
reconx = "reconx.__main__:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["reconx*"]
